Fixed some button issues with loadingtext and exported whole function
