+++ src/app.css
@@ -51,7 +51,8 @@ h6 {
   color: white;
 }
 
-.favCities {
+.favCities,
+.favCitiesBox {
   display: flex;
   justify-content: space-evenly;
   flex-flow: wrap;
@@ -91,7 +92,7 @@ h6 {
   display: flex;
   justify-content: center;
   align-items: flex-end;
-  margin-top: 3rem;
+  margin-top: 0.5rem;
 }
 
 @media only screen and (max-width: 600px) {
@@ -123,6 +124,13 @@ h6 {
   border: 2px solid limegreen;
 }
 
+.loadingContainer {
+  width: 500px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
 .loadingImg {
   width: 200px;
   height: 200px;
@@ -132,5 +140,7 @@ h6 {
 
 .loadingText {
   color: limegreen;
-  width: 80%;
+  width: 60%;
+  max-width: 500px;
+  margin: 1rem auto;
 }
